+++ frontend/app/src/components/history/events/HistoryEventsView.vue
@@ -12,6 +12,7 @@ import type { AddressData, BlockchainAccount } from '@/types/blockchain/accounts
 import type {
   AddTransactionHashPayload,
   HistoryEventRow,
+  PullEthBlockEventPayload,
   PullEvmTransactionPayload,
   RepullingTransactionPayload,
 } from '@/types/history/events';
@@ -125,8 +126,19 @@ const { decodingStatus } = storeToRefs(useHistoryStore());
 const { getAccountByAddress } = useBlockchainAccountsStore();
 const { fetchHistoryEvents } = useHistoryEvents();
 const { refreshTransactions } = useHistoryTransactions();
-const { fetchUndecodedTransactionsStatus, pullAndRedecodeTransactions, redecodeTransactions } = useHistoryTransactionDecoding();
-const { eventTaskLoading, processing, refreshing, sectionLoading, shouldFetchEventsRegularly } = useHistoryEventsStatus();
+const {
+  fetchUndecodedTransactionsStatus,
+  pullAndRecodeEthBlockEvents,
+  pullAndRedecodeTransactions,
+  redecodeTransactions,
+} = useHistoryTransactionDecoding();
+const {
+  anyEventsDecoding,
+  processing,
+  refreshing,
+  sectionLoading,
+  shouldFetchEventsRegularly,
+} = useHistoryEventsStatus();
 const historyEventMappings = useHistoryEventMappings();
 useHistoryEventsAutoFetch(shouldFetchEventsRegularly, fetchDataAndLocations);
 
@@ -362,6 +374,12 @@ async function fetchAndRedecodeEvents(data?: PullEvmTransactionPayload): Promise
     await forceRedecodeEvmEvents(data);
 }
 
+async function redecodeBlockEvents(data: PullEthBlockEventPayload): Promise<void> {
+  set(currentAction, 'decode');
+  await pullAndRecodeEthBlockEvents(data);
+  await fetchData();
+}
+
 function onShowDialog(type: 'decode' | 'protocol-refresh'): void {
   if (type === 'decode')
     set(decodingStatusDialogOpen, true);
@@ -415,7 +433,7 @@ watchImmediate(route, async (route) => {
   }
 });
 
-watch(eventTaskLoading, async (isLoading, wasLoading) => {
+watch(anyEventsDecoding, async (isLoading, wasLoading) => {
   if (!isLoading && wasLoading)
     await fetchDataAndLocations();
 });
@@ -453,7 +471,7 @@ onMounted(async () => {
       <HistoryEventsViewButtons
         v-model:open-decoding-dialog="decodingStatusDialogOpen"
         :processing="processing"
-        :loading="eventTaskLoading"
+        :loading="anyEventsDecoding"
         :include-evm-events="includes.evmEvents"
         @refresh="refresh(true, $event)"
         @show:form="showForm($event)"
@@ -530,6 +548,7 @@ onMounted(async () => {
           :highlighted-identifiers="highlightedIdentifiers"
           @show:form="showForm($event)"
           @refresh="fetchAndRedecodeEvents($event)"
+          @refresh:block-event="redecodeBlockEvents($event)"
           @set-page="setPage($event)"
         >
           <template #query-status="{ colspan }">
@@ -538,7 +557,6 @@ onMounted(async () => {
               :only-chains="onlyChains"
               :locations="locations"
               :decoding-status="decodingStatus"
-              :decoding="eventTaskLoading"
               :colspan="colspan"
               :loading="processing"
               @show:dialog="onShowDialog($event)"